// src/types/data.types.ts
  export interface Item {
    id: number;
    title: string;
    name?: string;
    category?: string;
    newPrice: number;
    waight?: string;
    sale?: string;
    image: string;
    imageTwo?: string;
    date?: string;
    oldPrice: number;
    href?: string;
    weight?: string;
    rating?: any;
    status?: string;
    location?: string;
    brand?: string;
    sku?: number;
    quantity: number;
    color?: string;
    tags?: string;
    wishlist?: string;
    color1?: string;
    color2?: string;
    color3?: string;
    size1?: string;
    size2?: string;
  }